# backend/docs/openapi-governance.yaml
openapi: 3.1.0
info:
  title: DictionaryRoot Governed Platform API
  version: 0.2.0
  description: Authentication, account, governance, workflow, moderation, and audit additions to SourceRoot.
servers:
  - url: http://localhost:3000/api/v1
paths:
...
  /account:
    get:
      summary: Read the authenticated account
      responses: { '200': { description: Account context }, '401': { description: Sign-in required } }
  /account/export:
    get:
      summary: Download account data as JSON
      responses: { '200': { description: Account export }, '401': { description: Sign-in required } }
  /account/invitations/accept:
    post:
      summary: Accept an organization invitation matching a verified account email
      responses: { '200': { description: Invitation accepted }, '403': { description: Email mismatch } }
...
  /account/sessions:
    get:
      summary: List active browser sessions
      responses: { '200': { description: Session list }, '401': { description: Sign-in required } }
...
components:
  securitySchemes:
    dictionaryRootSession:
      type: apiKey
      in: cookie
      name: dr_session
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token